requirements.txt
# Core Dependencies
google-api-python-client>=2.100.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Network Analysis
networkx>=3.1
python-louvain>=0.16

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1

# Optional: Advanced community detection
# cdlib>=0.2.7
# python-igraph>=0.10.6

# Optional: Feature extraction
# gensim>=4.3.1
# sentence-transformers>=2.2.2
# scikit-learn>=1.3.0

